Correcciones varias
Mejoras en el cierre de sesión si se abre pestañas en el navegador, correcciones en personal, se cambia int a double en cantidades y stock en todas las tablas, se corige error en cambiar contraseña

+++ Controllers/AjusteController.cs
@@ -150,8 +150,8 @@ namespace Gestion_Compras.Controllers
                         usuarioId = uid;
                 }
                 string itemCodigo = ajusteData.ItemCodigo;
-                int stockIni = ajusteData.StockIni;
-                int stockReal = ajusteData.StockReal;
+                double stockIni = ajusteData.StockIni;
+                double stockReal = ajusteData.StockReal;
                 string observaciones = ajusteData.Observaciones ?? "";
 
                 if (string.IsNullOrEmpty(itemCodigo))
